+++ app/routes/login.tsx
@@ -10,12 +10,13 @@ import {
 } from "remix";
 import { db } from "~/utils/db.server";
 import { invariant } from "~/utils/invariant";
+import { guaranteePermission, Permissions } from "~/utils/permission";
 import { passwordScheme, usernameScheme } from "~/utils/scheme";
 import { commitSession } from "~/utils/sessions";
 import { parseRedirectPathname } from "~/utils/tools";
 
 export const meta: MetaFunction = () => ({
-  title: "Login",
+  title: "登录 - HITwh OJ",
 });
 
 export const action: ActionFunction = async ({ request }) => {
@@ -36,6 +37,11 @@ export const action: ActionFunction = async ({ request }) => {
   if (user.password !== password) {
     return new Response("Password is incorrect", { status: 400 });
   }
+
+  // TODO: 看起来有点奇怪
+  await guaranteePermission(request, Permissions.User.Session.Create, {
+    uid: user.uid,
+  });
 
   return redirect(parseRedirectPathname(form.get("redirect")), {
     headers: {
@@ -76,3 +82,6 @@ export default function Login() {
     </>
   );
 }
+
+export { ErrorBoundary } from "~/src/ErrorBoundary";
+export { CatchBoundary } from "~/src/CatchBoundary";