+++ app/routes/problem/$pid/submit.tsx
@@ -1,32 +1,54 @@
-import { ActionFunction, Form, LoaderFunction, redirect } from "remix";
+import {
+  ActionFunction,
+  Form,
+  json,
+  LoaderFunction,
+  MetaFunction,
+  redirect,
+} from "remix";
 import { db } from "~/utils/db.server";
 import { s3 } from "~/utils/s3.server";
 import { invariant } from "~/utils/invariant";
 import { codeScheme, idScheme } from "~/utils/scheme";
-import { findSessionUid } from "~/utils/sessions";
 import { judge } from "~/utils/judge.server";
 import { Button, Input, Space, Select } from "@arco-design/web-react";
 import { useState } from "react";
 const TextArea = Input.TextArea;
+import { guaranteePermission, Permissions } from "~/utils/permission";
+import { Problem } from "@prisma/client";
 
-export const loader: LoaderFunction = async ({ request }) => {
-  const self = await findSessionUid(request);
-
-  if (!self) {
-    return redirect(`/login?redirect=${new URL(request.url).pathname}`);
-  }
-
-  return null;
+type LoaderData = {
+  problem: Pick<Problem, "title">;
 };
 
-export const action: ActionFunction = async ({ params, request }) => {
-  const uid = await findSessionUid(request);
+export const loader: LoaderFunction = async ({ request, params }) => {
+  const pid = invariant(idScheme.safeParse(params.pid), { status: 404 });
 
-  if (!uid) {
-    return redirect(`/login?redirect=${new URL(request.url).pathname}`);
+  await guaranteePermission(request, Permissions.Problem.Submit, { pid });
+
+  const problem = await db.problem.findUnique({
+    where: { pid },
+    select: { title: true },
+  });
+
+  if (!problem) {
+    throw new Response("Problem not found", { status: 404 });
   }
 
+  return json({ problem });
+};
+
+export const meta: MetaFunction = ({ data }: { data?: LoaderData }) => ({
+  title: `提交题目: ${data?.problem.title} - HITwh OJ`,
+});
+
+export const action: ActionFunction = async ({ request, params }) => {
   const pid = invariant(idScheme.safeParse(params.pid), { status: 404 });
+
+  // 检查提交权限，用户必定已经登录
+  const uid = (await guaranteePermission(request, Permissions.Problem.Submit, {
+    pid,
+  }))!;
 
   const form = await request.formData();
   const code = invariant(codeScheme.safeParse(form.get("code")));
@@ -87,3 +109,6 @@ export default function ProblemSubmit() {
     </>
   );
 }
+
+export { ErrorBoundary } from "~/src/ErrorBoundary";
+export { CatchBoundary } from "~/src/CatchBoundary";