+++ app/routes/register.tsx
@@ -10,12 +10,13 @@ import {
 } from "remix";
 import { db } from "~/utils/db.server";
 import { invariant } from "~/utils/invariant";
+import { guaranteePermission, Permissions } from "~/utils/permission";
 import { passwordScheme, usernameScheme } from "~/utils/scheme";
 import { commitSession } from "~/utils/sessions";
 import { parseRedirectPathname } from "~/utils/tools";
 
 export const meta: MetaFunction = () => ({
-  title: "Register",
+  title: "注册 - HITwh OJ",
 });
 
 export const action: ActionFunction = async ({ request }) => {
@@ -28,6 +29,8 @@ export const action: ActionFunction = async ({ request }) => {
   if (password !== password2) {
     return new Response("Passwords do not match", { status: 400 });
   }
+
+  await guaranteePermission(request, Permissions.User.Session.Create);
 
   if (await db.user.findUnique({ where: { username } })) {
     return new Response("username is already taken", { status: 400 });
@@ -84,3 +87,6 @@ export default function Register() {
     </>
   );
 }
+
+export { ErrorBoundary } from "~/src/ErrorBoundary";
+export { CatchBoundary } from "~/src/CatchBoundary";