+++ app/routes/user/$uid/files.tsx
@@ -1,12 +1,13 @@
 import { Button, Space, Switch, Table } from "@arco-design/web-react";
 import { ColumnProps } from "@arco-design/web-react/es/Table";
 import { IconDelete, IconUpload } from "@arco-design/web-react/icon";
-import { File as UserFile } from "@prisma/client";
+import { File as UserFile, User } from "@prisma/client";
 import { useRef } from "react";
 import {
   ActionFunction,
   json,
   LoaderFunction,
+  MetaFunction,
   unstable_parseMultipartFormData,
   useFetcher,
   useLoaderData,
@@ -14,19 +15,26 @@ import {
 import { db } from "~/utils/db.server";
 import { createUserFile, removeFile } from "~/utils/files";
 import { invariant } from "~/utils/invariant";
+import { guaranteePermission, Permissions } from "~/utils/permission";
 import { idScheme, uuidScheme } from "~/utils/scheme";
 import { uploadHandler } from "~/utils/uploadHandler";
 
 type LoaderData = {
-  files: UserFile[];
+  user: Pick<User, "nickname" | "username"> & {
+    createdFiles: UserFile[];
+  };
 };
 
-export const loader: LoaderFunction = async ({ params }) => {
+export const loader: LoaderFunction = async ({ request, params }) => {
   const uid = invariant(idScheme.safeParse(params.uid), { status: 404 });
+
+  await guaranteePermission(request, Permissions.User.File.View, { uid });
 
   const user = await db.user.findUnique({
     where: { uid },
     select: {
+      username: true,
+      nickname: true,
       createdFiles: {
         orderBy: [{ createdAt: "desc" }, { filename: "asc" }],
       },
@@ -37,8 +45,12 @@ export const loader: LoaderFunction = async ({ params }) => {
     throw new Response("User not found", { status: 404 });
   }
 
-  return json({ files: user.createdFiles });
+  return json({ user });
 };
+
+export const meta: MetaFunction = ({ data }: { data?: LoaderData }) => ({
+  title: `用户文件: ${data?.user.nickname || data?.user.username} - HITwh OJ`,
+});
 
 enum ActionType {
   UploadFile = "uploadFile",
@@ -54,6 +66,8 @@ export const action: ActionFunction = async ({ request, params }) => {
 
   switch (_action) {
     case ActionType.UploadFile: {
+      await guaranteePermission(request, Permissions.User.File.Create, { uid });
+
       const files = form
         .getAll("file")
         .filter((file): file is File => file instanceof File);
@@ -68,9 +82,9 @@ export const action: ActionFunction = async ({ request, params }) => {
     }
 
     case ActionType.RemoveFile: {
-      const fid = invariant(uuidScheme.safeParse(form.get("fid")), {
-        status: 400,
-      });
+      await guaranteePermission(request, Permissions.User.File.Delete, { uid });
+
+      const fid = invariant(uuidScheme.safeParse(form.get("fid")));
 
       await removeFile(fid);
 
@@ -78,9 +92,9 @@ export const action: ActionFunction = async ({ request, params }) => {
     }
 
     case ActionType.ModifyPrivacy: {
-      const fid = invariant(uuidScheme.safeParse(form.get("fid")), {
-        status: 400,
-      });
+      await guaranteePermission(request, Permissions.User.File.Update, { uid });
+
+      const fid = invariant(uuidScheme.safeParse(form.get("fid")));
 
       const file = await db.file.findUnique({
         where: { fid },
@@ -158,14 +172,13 @@ function UserFileRemoveButton({ file }: { file: UserFile }) {
     <fetcher.Form method="post">
       <input type="hidden" name="fid" value={file.fid} />
       <Button
-        type="primary"
+        type="text"
         status="danger"
         htmlType="submit"
         name="_action"
         value={ActionType.RemoveFile}
         loading={isDeleting}
         icon={<IconDelete />}
-        size="small"
       />
     </fetcher.Form>
   );
@@ -175,13 +188,13 @@ const columns: ColumnProps<UserFile>[] = [
   {
     title: "文件名",
     dataIndex: "filename",
-    sorter: (a, b) =>
+    sorter: (a: UserFile, b: UserFile) =>
       a.filename > b.filename ? 1 : a.filename < b.filename ? -1 : 0,
   },
   {
     title: "文件大小",
     dataIndex: "filesize",
-    sorter: (a, b) => a.filesize - b.filesize,
+    sorter: (a: UserFile, b: UserFile) => a.filesize - b.filesize,
   },
   {
     title: "文件类型",
@@ -215,7 +228,9 @@ function UserFileList({ files }: { files: UserFile[] }) {
 }
 
 export default function UserFiles() {
-  const { files } = useLoaderData<LoaderData>();
+  const {
+    user: { createdFiles: files },
+  } = useLoaderData<LoaderData>();
 
   return (
     <Space direction="vertical" size="medium" style={{ display: "flex" }}>
@@ -224,3 +239,6 @@ export default function UserFiles() {
     </Space>
   );
 }
+
+export { ErrorBoundary } from "~/src/ErrorBoundary";
+export { CatchBoundary } from "~/src/CatchBoundary";