+++ app/routes/user/$uid/index.tsx
@@ -1,16 +1,19 @@
 import { Descriptions } from "@arco-design/web-react";
 import { User } from "@prisma/client";
-import { json, LoaderFunction, useLoaderData } from "remix";
+import { json, LoaderFunction, MetaFunction, useLoaderData } from "remix";
 import { db } from "~/utils/db.server";
 import { invariant } from "~/utils/invariant";
+import { guaranteePermission, Permissions } from "~/utils/permission";
 import { idScheme } from "~/utils/scheme";
 
 type LoaderData = {
   user: Pick<User, "uid" | "username" | "nickname" | "email" | "password">;
 };
 
-export const loader: LoaderFunction = async ({ params }) => {
+export const loader: LoaderFunction = async ({ request, params }) => {
   const uid = invariant(idScheme.safeParse(params.uid), { status: 404 });
+
+  await guaranteePermission(request, Permissions.User.Profile.View, { uid });
 
   const user = await db.user.findUnique({
     where: { uid },
@@ -29,6 +32,10 @@ export const loader: LoaderFunction = async ({ params }) => {
 
   return json({ user });
 };
+
+export const meta: MetaFunction = ({ data }: { data?: LoaderData }) => ({
+  title: `用户: ${data?.user.nickname || data?.user.username} - HITwh OJ`,
+});
 
 export default function Profile() {
   const { user } = useLoaderData<LoaderData>();
@@ -61,3 +68,6 @@ export default function Profile() {
     />
   );
 }
+
+export { ErrorBoundary } from "~/src/ErrorBoundary";
+export { CatchBoundary } from "~/src/CatchBoundary";