+++ app/utils/permission.ts
@@ -1,6 +1,6 @@
 import { SystemUserRole } from "@prisma/client";
-
 import { db } from "./db.server";
+import { findSessionUid } from "./sessions";
 
 let count = 0n;
 
@@ -12,67 +12,52 @@ export const Permissions = {
   ALL: -1n,
   DEFAULT: 0n,
 
-  USER_SESSION_LIST: perm(),
-  USER_SESSION_REMOVE: perm(),
-  USER_SESSION_SIGN_UP: perm(),
-  USER_SESSION_SIGN_IN: perm(),
+  User: {
+    Register: perm(),
+    Session: {
+      View: perm(),
+      Create: perm(),
+      Delete: perm(),
+    },
+    Profile: {
+      View: perm(),
+      Update: perm(),
+    },
+    File: {
+      View: perm(),
+      Create: perm(),
+      Update: perm(),
+      Delete: perm(),
+      DownloadPublic: perm(),
+      DownloadPrivate: perm(),
+    },
+  },
 
-  USER_PROFILE_GET: perm(),
-  USER_PROFILE_MODIFY: perm(),
-
-  USER_FILE_LIST: perm(),
-  USER_FILE_UPLOAD: perm(),
-  USER_FILE_UNLINK: perm(),
-  USER_FILE_DOWNLOAD: perm(),
-
-  PROBLEM_GET: perm(),
-  PROBLEM_LIST: perm(),
-  PROBLEM_CREATE: perm(),
-  PROBLEM_MODIFY: perm(),
-  PROBLEM_REMOVE: perm(),
-  PROBLEM_SEARCH: perm(),
-
-  PROBLEM_SUBMIT: perm(),
-
-  PROBLEM_DATA_UPLOAD: perm(),
-  PROBLEM_DATA_REMOVE: perm(),
-  PROBLEM_DATA_DOWNLOAD: perm(),
-  PROBLEM_FILE_UPLOAD: perm(),
-  PROBLEM_FILE_REMOVE: perm(),
-  PROBLEM_FILE_DOWNLOAD: perm(),
-
-  PROBLEM_SET_GET: perm(),
-  PROBLEM_SET_LIST: perm(),
-  PROBLEM_SET_CREATE: perm(),
-  PROBLEM_SET_MODIFY: perm(),
-  PROBLEM_SET_REMOVE: perm(),
-  PROBLEM_SET_SEARCH: perm(),
-
-  CONTEST_GET: perm(),
-  CONTEST_LIST: perm(),
-  CONTEST_CREATE: perm(),
-  CONTEST_MODIFY: perm(),
-  CONTEST_REMOVE: perm(),
-  CONTEST_SEARCH: perm(),
-
-  CONTEST_PROBLEM_GET: perm(),
-  CONTEST_PROBLEM_LIST: perm(),
-  CONTEST_PROBLEM_SUBMIT: perm(),
-
-  RECORD_GET: perm(),
-  RECORD_LIST: perm(),
-  RECORD_VIEW_CODE: perm(),
+  Problem: {
+    Create: perm(),
+    Update: perm(),
+    Delete: perm(),
+    ViewPublic: perm(),
+    ViewPrivate: perm(),
+    Submit: perm(),
+    Data: {
+      View: perm(),
+      Create: perm(),
+      Download: perm(),
+      Delete: perm(),
+    },
+    File: {
+      View: perm(),
+      Create: perm(),
+      Download: perm(),
+      Delete: perm(),
+    },
+  },
 };
 
 enum CustomUserRole {
   UserSelf = "UserSelf",
-
   ProblemAdmin = "ProblemAdmin",
-
-  ContestAdmin = "ContestAdmin",
-  ContestMod = "ContestMod",
-  ContestJury = "ContestJury",
-  ContestAttendee = "ContestAttendee",
 }
 
 type UserRole = SystemUserRole | CustomUserRole;
@@ -81,158 +66,134 @@ const PermissionDict: Record<UserRole, bigint> = {
   // 超级用户
   [SystemUserRole.Su]: Permissions.ALL,
 
-  // 网站管理员比普通用户多的权限
+  // 网站管理员
   [SystemUserRole.Admin]:
     Permissions.DEFAULT |
-    Permissions.USER_SESSION_LIST |
-    Permissions.USER_SESSION_REMOVE |
-    Permissions.USER_PROFILE_GET |
-    Permissions.USER_PROFILE_MODIFY |
-    Permissions.USER_FILE_LIST |
-    Permissions.USER_FILE_UNLINK |
-    Permissions.USER_FILE_DOWNLOAD |
-    Permissions.PROBLEM_GET |
-    Permissions.PROBLEM_LIST |
-    Permissions.PROBLEM_CREATE |
-    Permissions.PROBLEM_MODIFY |
-    Permissions.PROBLEM_REMOVE |
-    Permissions.PROBLEM_SEARCH |
-    Permissions.PROBLEM_DATA_UPLOAD |
-    Permissions.PROBLEM_DATA_REMOVE |
-    Permissions.PROBLEM_DATA_DOWNLOAD |
-    Permissions.PROBLEM_FILE_UPLOAD |
-    Permissions.PROBLEM_FILE_REMOVE |
-    Permissions.PROBLEM_FILE_DOWNLOAD |
-    Permissions.PROBLEM_SET_GET |
-    Permissions.PROBLEM_SET_LIST |
-    Permissions.PROBLEM_SET_CREATE |
-    Permissions.PROBLEM_SET_MODIFY |
-    Permissions.PROBLEM_SET_REMOVE |
-    Permissions.PROBLEM_SET_SEARCH |
-    Permissions.CONTEST_GET |
-    Permissions.CONTEST_LIST |
-    Permissions.CONTEST_CREATE |
-    Permissions.CONTEST_MODIFY |
-    Permissions.CONTEST_REMOVE |
-    Permissions.CONTEST_SEARCH |
-    Permissions.CONTEST_PROBLEM_GET |
-    Permissions.CONTEST_PROBLEM_LIST |
-    Permissions.RECORD_GET |
-    Permissions.RECORD_LIST |
-    Permissions.RECORD_VIEW_CODE,
+    Permissions.User.Profile.View |
+    Permissions.User.Profile.Update |
+    Permissions.User.File.View |
+    Permissions.User.File.Update |
+    Permissions.User.File.Delete |
+    Permissions.User.File.DownloadPublic |
+    Permissions.User.File.DownloadPrivate |
+    Permissions.Problem.Create |
+    Permissions.Problem.Update |
+    Permissions.Problem.Delete |
+    Permissions.Problem.ViewPublic |
+    Permissions.Problem.ViewPrivate |
+    Permissions.Problem.Data.View |
+    Permissions.Problem.Data.Create |
+    Permissions.Problem.Data.Download |
+    Permissions.Problem.Data.Delete |
+    Permissions.Problem.File.View |
+    Permissions.Problem.File.Create |
+    Permissions.Problem.File.Download |
+    Permissions.Problem.File.Delete |
+    Permissions.Problem.Submit,
 
   // 普通用户的权限
   [SystemUserRole.User]:
     Permissions.DEFAULT |
-    Permissions.USER_PROFILE_GET |
-    Permissions.USER_FILE_DOWNLOAD |
-    Permissions.PROBLEM_GET |
-    Permissions.PROBLEM_LIST |
-    Permissions.PROBLEM_SEARCH |
-    Permissions.PROBLEM_FILE_DOWNLOAD |
-    Permissions.PROBLEM_SET_GET |
-    Permissions.PROBLEM_SET_LIST |
-    Permissions.PROBLEM_SET_SEARCH |
-    Permissions.CONTEST_GET |
-    Permissions.CONTEST_LIST |
-    Permissions.CONTEST_SEARCH |
-    Permissions.RECORD_GET |
-    Permissions.RECORD_LIST,
+    Permissions.User.Profile.View |
+    Permissions.User.File.DownloadPublic |
+    Permissions.Problem.ViewPublic |
+    Permissions.Problem.File.View |
+    Permissions.Problem.File.Download |
+    Permissions.Problem.Submit,
 
   // 访客权限
   [SystemUserRole.Guest]:
     Permissions.DEFAULT |
-    Permissions.USER_SESSION_SIGN_UP |
-    Permissions.USER_SESSION_SIGN_IN |
-    Permissions.USER_PROFILE_GET |
-    Permissions.USER_FILE_DOWNLOAD |
-    Permissions.PROBLEM_GET |
-    Permissions.PROBLEM_LIST |
-    Permissions.PROBLEM_SEARCH |
-    Permissions.PROBLEM_FILE_DOWNLOAD |
-    Permissions.PROBLEM_SET_GET |
-    Permissions.PROBLEM_SET_LIST |
-    Permissions.PROBLEM_SET_SEARCH |
-    Permissions.CONTEST_GET |
-    Permissions.CONTEST_LIST |
-    Permissions.CONTEST_SEARCH |
-    Permissions.RECORD_GET |
-    Permissions.RECORD_LIST,
+    Permissions.User.Register |
+    Permissions.User.Session.Create |
+    Permissions.User.Profile.View |
+    Permissions.User.File.DownloadPublic |
+    Permissions.Problem.ViewPublic |
+    Permissions.Problem.File.View |
+    Permissions.Problem.File.Download,
 
-  // 题目管理员
-  [CustomUserRole.ProblemAdmin]:
-    Permissions.DEFAULT |
-    Permissions.PROBLEM_MODIFY |
-    Permissions.PROBLEM_REMOVE |
-    Permissions.PROBLEM_DATA_UPLOAD |
-    Permissions.PROBLEM_DATA_REMOVE |
-    Permissions.PROBLEM_DATA_DOWNLOAD |
-    Permissions.PROBLEM_FILE_UPLOAD |
-    Permissions.PROBLEM_FILE_REMOVE |
-    Permissions.PROBLEM_FILE_DOWNLOAD |
-    Permissions.RECORD_VIEW_CODE,
-
-  // 比赛管理员
-  [CustomUserRole.ContestAdmin]:
-    Permissions.DEFAULT |
-    Permissions.CONTEST_MODIFY |
-    Permissions.CONTEST_REMOVE |
-    Permissions.CONTEST_PROBLEM_GET |
-    Permissions.CONTEST_PROBLEM_LIST |
-    Permissions.RECORD_VIEW_CODE,
-
-  // 比赛负责人
-  [CustomUserRole.ContestMod]:
-    Permissions.DEFAULT |
-    Permissions.CONTEST_MODIFY |
-    Permissions.CONTEST_REMOVE |
-    Permissions.CONTEST_PROBLEM_GET |
-    Permissions.CONTEST_PROBLEM_LIST |
-    Permissions.RECORD_VIEW_CODE,
-
-  // 比赛裁判（负责赛事答疑）
-  [CustomUserRole.ContestJury]:
-    Permissions.DEFAULT |
-    Permissions.CONTEST_PROBLEM_GET |
-    Permissions.CONTEST_PROBLEM_LIST,
-
-  // 比赛参赛者
-  [CustomUserRole.ContestAttendee]:
-    Permissions.DEFAULT |
-    Permissions.CONTEST_PROBLEM_GET |
-    Permissions.CONTEST_PROBLEM_LIST |
-    Permissions.CONTEST_PROBLEM_SUBMIT,
-
-  // 已登录用户
+  // 用户自己
   [CustomUserRole.UserSelf]:
     Permissions.DEFAULT |
-    Permissions.USER_FILE_LIST |
-    Permissions.USER_FILE_UPLOAD |
-    Permissions.USER_FILE_UNLINK |
-    Permissions.USER_FILE_DOWNLOAD |
-    Permissions.PROBLEM_SUBMIT,
+    Permissions.User.Session.View |
+    Permissions.User.Session.Delete |
+    Permissions.User.Profile.View |
+    Permissions.User.Profile.Update |
+    Permissions.User.File.View |
+    Permissions.User.File.Create |
+    Permissions.User.File.Update |
+    Permissions.User.File.Delete |
+    Permissions.User.File.DownloadPublic |
+    Permissions.User.File.DownloadPrivate,
+
+  // 题目创建者（管理员）
+  [CustomUserRole.ProblemAdmin]:
+    Permissions.DEFAULT |
+    Permissions.Problem.Create |
+    Permissions.Problem.Update |
+    Permissions.Problem.Delete |
+    Permissions.Problem.ViewPublic |
+    Permissions.Problem.ViewPrivate |
+    Permissions.Problem.Data.View |
+    Permissions.Problem.Data.Create |
+    Permissions.Problem.Data.Download |
+    Permissions.Problem.Data.Delete |
+    Permissions.Problem.File.View |
+    Permissions.Problem.File.Create |
+    Permissions.Problem.File.Download |
+    Permissions.Problem.File.Delete,
 };
 
-// FIXME: 需要重新设计
+export type GuaranteePermissionInfo = {
+  /**
+   * 目标用户 ID
+   */
+  uid?: number;
+  /**
+   * 目标题目 ID
+   */
+  pid?: number;
+  /**
+   * 目标比赛 ID
+   */
+  cid?: number;
+  /**
+   * 目标评测 ID
+   */
+  rid?: number;
+  /**
+   * 目标文件 ID
+   */
+  fid?: string;
+};
+
+/**
+ * 检查当前登录的用户是否有相应的权限
+ *
+ * 如果有，则返回当前登录的用户 ID，否则抛出 403 Response
+ *
+ * @param request 当前请求
+ * @param permission 权限
+ * @param info 附加的限定信息
+ */
 export async function guaranteePermission(
+  request: Request,
   permission: bigint,
-  config?: {
-    uid: number;
-    pid: number;
-    cid: number;
-    rid: number;
-  }
-): Promise<void> {
+  info?: GuaranteePermissionInfo
+) {
   const roles: UserRole[] = [];
 
-  if (config?.uid) {
+  const self = await findSessionUid(request);
+
+  // 用户自身系统权限
+  if (self) {
     const user = await db.user.findUnique({
-      where: { uid: config.uid },
+      where: { uid: self },
       select: { role: true },
     });
 
     if (!user) {
-      throw new Response("Permission: User not found", { status: 404 });
+      throw new Response("User not found", { status: 500 });
     }
 
     roles.push(user.role);
@@ -240,14 +201,22 @@ export async function guaranteePermission(
     roles.push(SystemUserRole.Guest);
   }
 
-  if (config?.uid && config?.pid) {
-    roles.push(...(await getProblemRoles(config.uid, config.pid)));
+  // 用户对自身操作权限
+  if (self && info?.uid === self) {
+    roles.push(CustomUserRole.UserSelf);
   }
-  if (config?.uid && config?.cid) {
-    roles.push(...(await getContestRoles(config.uid, config.cid)));
-  }
-  if (config?.uid && config?.rid) {
-    roles.push(...(await getRecordRoles(config.uid, config.rid)));
+
+  // 题目管理员的权限
+  if (self && info?.pid) {
+    // TODO: 只检查了是否是题目的创建者，没有判断其他可能的角色
+    const problem = await db.problem.findUnique({
+      where: { pid: info.pid },
+      select: { uid: true },
+    });
+
+    if (problem && problem.uid === self) {
+      roles.push(CustomUserRole.ProblemAdmin);
+    }
   }
 
   const userPermission = roles
@@ -255,79 +224,15 @@ export async function guaranteePermission(
     .reduce((perm, rolePerm) => perm | rolePerm, 0n);
 
   if ((userPermission & permission) !== permission) {
+    // 没有权限
+    if (!self) {
+      // 如果没有登录，则返回一个 401 错误
+      throw new Response("Unauthorized", { status: 401 });
+    }
+
+    // 否则抛出 403
     throw new Response("Permission denied", { status: 403 });
   }
-}
 
-export async function getProblemRoles(uid: number, pid: number) {
-  const problem = await db.problem.findUnique({
-    where: { pid },
-    select: { uid: true },
-  });
-
-  if (!problem) {
-    throw new Response("Permissions: Problem not found", { status: 404 });
-  }
-
-  const roles: UserRole[] = [];
-
-  if (problem.uid === uid) {
-    roles.push(CustomUserRole.ProblemAdmin);
-  }
-
-  return roles;
-}
-
-export async function getContestRoles(uid: number, cid: number) {
-  const contest = await db.contest.findUnique({
-    where: { cid },
-    select: {
-      user: { select: { uid: true } },
-      mods: { select: { uid: true } },
-      juries: { select: { uid: true } },
-      attendees: { select: { uid: true } },
-    },
-  });
-
-  if (!contest) {
-    throw new Response("Permissions: Contest not found", { status: 404 });
-  }
-
-  const roles: UserRole[] = [];
-
-  if (contest.user.uid === uid) {
-    roles.push(CustomUserRole.ContestAdmin);
-  }
-  if (contest.mods.map(({ uid }) => uid).includes(uid)) {
-    roles.push(CustomUserRole.ContestMod);
-  }
-  if (contest.juries.map(({ uid }) => uid).includes(uid)) {
-    roles.push(CustomUserRole.ContestJury);
-  }
-  if (contest.attendees.map(({ uid }) => uid).includes(uid)) {
-    roles.push(CustomUserRole.ContestAttendee);
-  }
-
-  return roles;
-}
-
-export async function getRecordRoles(uid: number, rid: number) {
-  const record = await db.record.findUnique({
-    where: { rid },
-    select: { uid: true },
-  });
-
-  if (!record) {
-    throw new Response("Permissions: Record not found", { status: 404 });
-  }
-
-  // avoid infomation leak
-  // also sprach eryi
-  const roles: UserRole[] = [];
-
-  if (record.uid === uid) {
-    roles.push(CustomUserRole.UserSelf);
-  }
-
-  return roles;
+  return self;
 }