+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- feat: add problem permissions
+- feat: added user permissions
 - fix: use `defaultValue` in `<select>` instead of setting `selected` in `<option>`
 - feat: add redirect searchParams in `register.tsx`
 - Added CHANGELOG.md